+++ core/src/main/java/com/github/games647/changeskin/core/MojangSkinApi.java
@@ -1,8 +1,11 @@
 package com.github.games647.changeskin.core;
 
+import com.github.games647.changeskin.core.model.ApiPropertiesModel;
+import com.github.games647.changeskin.core.model.McApiProfile;
 import com.github.games647.changeskin.core.model.PlayerProfile;
-import com.github.games647.changeskin.core.model.PropertiesModel;
-import com.github.games647.changeskin.core.model.TexturesModel;
+import com.github.games647.changeskin.core.model.RawPropertiesModel;
+import com.github.games647.changeskin.core.model.mojang.PropertiesModel;
+import com.github.games647.changeskin.core.model.mojang.TexturesModel;
 import com.google.common.io.CharStreams;
 import com.google.common.io.Closeables;
 import com.google.gson.Gson;
@@ -21,6 +24,9 @@ import java.util.logging.Logger;
 public class MojangSkinApi {
 
     private static final String SKIN_URL = "https://sessionserver.mojang.com/session/minecraft/profile/";
+    private static final String MCAPI_SKIN_URL = "https://mcapi.de/api/user/";
+
+
     private static final String UUID_URL = "https://api.mojang.com/users/profiles/minecraft/";
 //    private static final String MCAPI_UUID_URL = "https://mcapi.de/api/user/";
 //    private static final String MCAPI_UUID_URL = "https://us.mc-api.net/v3/uuid/";
@@ -33,15 +39,17 @@ public class MojangSkinApi {
     private final Gson gson = new Gson();
 
     private final ConcurrentMap<Object, Object> requests;
-    private final int rateLimit;
     private final Logger logger;
+    private final int rateLimit;
+    private final boolean mojangDownload;
 
     private long lastRateLimit;
 
-    public MojangSkinApi(ConcurrentMap<Object, Object> requests, int rateLimit, Logger logger) {
+    public MojangSkinApi(ConcurrentMap<Object, Object> requests, Logger logger, int rateLimit, boolean mojangDownload) {
         this.requests = requests;
         this.rateLimit = rateLimit;
         this.logger = logger;
+        this.mojangDownload = mojangDownload;
     }
 
     public UUID getUUID(String playerName) throws NotPremiumException, RateLimitException {
@@ -113,6 +121,10 @@ public class MojangSkinApi {
     }
 
     public SkinData downloadSkin(UUID ownerUUID) {
+        if (mojangDownload) {
+            return downloadSkinFromApi(ownerUUID);
+        }
+
         //unsigned is needed in order to receive the signature
         String uuidString = ownerUUID.toString().replace("-", "") + "?unsigned=false";
         try {
@@ -127,6 +139,39 @@ public class MojangSkinApi {
                 PropertiesModel[] properties = texturesModel.getProperties();
                 if (properties != null && properties.length > 0) {
                     PropertiesModel propertiesModel = properties[0];
+
+                    //base64 encoded skin data
+                    String encodedSkin = propertiesModel.getValue();
+                    String signature = propertiesModel.getSignature();
+
+                    SkinData skinData = new SkinData(encodedSkin, signature);
+                    return skinData;
+                }
+            }
+        } catch (IOException ioException) {
+            logger.log(Level.SEVERE, "Tried downloading skin data from Mojang", ioException);
+        } catch (JsonParseException parseException) {
+            logger.log(Level.SEVERE, "Tried parsing json from Mojang", parseException);
+        }
+
+        return null;
+    }
+
+    public SkinData downloadSkinFromApi(UUID ownerUUID) {
+        //unsigned is needed in order to receive the signature
+        String uuidStrip = ownerUUID.toString().replace("-", "");
+        try {
+            HttpURLConnection httpConnection = (HttpURLConnection) new URL(MCAPI_SKIN_URL + uuidStrip).openConnection();
+            httpConnection.addRequestProperty("Content-Type", "application/json");
+
+            BufferedReader reader = new BufferedReader(new InputStreamReader(httpConnection.getInputStream()));
+            String line = reader.readLine();
+            if (line != null && !line.equals("null")) {
+                McApiProfile profile = gson.fromJson(line, McApiProfile.class);
+
+                ApiPropertiesModel properties = profile.getProperties();
+                if (properties != null && properties.getRaw().length > 0) {
+                    RawPropertiesModel propertiesModel = properties.getRaw()[0];
 
                     //base64 encoded skin data
                     String encodedSkin = propertiesModel.getValue();