+++ core/src/main/java/com/github/games647/changeskin/core/model/ApiPropertiesModel.java
@@ -1,0 +1,10 @@
+package com.github.games647.changeskin.core.model;
+
+public class ApiPropertiesModel {
+
+    private RawPropertiesModel[] raw;
+
+    public RawPropertiesModel[] getRaw() {
+        return raw;
+    }
+}