+++ core/src/main/java/com/github/games647/changeskin/core/model/McApiProfile.java
@@ -1,10 +1,11 @@
 package com.github.games647.changeskin.core.model;
 
+//https://mcapi.de/api/user/0aaa2c13922a411bb6559b8c08404695
 public class McApiProfile {
 
-    private PlayerProfile[] profiles;
+    private ApiPropertiesModel properties;
 
-    public PlayerProfile[] getProfiles() {
-        return profiles;
+    public ApiPropertiesModel getProperties() {
+        return properties;
     }
 }