+++ core/src/main/java/com/github/games647/changeskin/core/model/RawPropertiesModel.java
@@ -1,0 +1,20 @@
+package com.github.games647.changeskin.core.model;
+
+public class RawPropertiesModel {
+
+    private String signature;
+    private String value;
+    private String name = "textures";
+
+    public String getSignature() {
+        return signature;
+    }
+
+    public String getValue() {
+        return value;
+    }
+
+    public String getName() {
+        return name;
+    }
+}